+++ tests/conftest.py
@@ -8,7 +8,7 @@ def pytest_addoption(parser):
     parser.addoption('--port'     ,action="store" ,default=8020)
     parser.addoption('--ip'       ,action="store" ,default="localhost")
     parser.addoption('--address'  ,action="store" ,default='')
-    parser.addoption('--sleep'    ,action="store" ,default=1)
+    parser.addoption('--sleep'    ,action="store" ,default=0.3)
     parser.addoption('--headless' ,action='store_true',default=False)
 
     parser.addoption("--slow", action="store_true", default=False, help="run slow tests")
@@ -22,7 +22,7 @@ def pytest_configure(config):
     config.addinivalue_line("markers", "slow: mark test as slow to run")
     config.addinivalue_line("markers", "opt: mark test as optional")
     config.addinivalue_line("markers", "old: mark test as old")
-    for i in range(1,3):
+    for i in range(1,5):
         config.addinivalue_line("markers", "lvl%d: test level %d" %(i,i))
     config.addinivalue_line("markers", "new: mark test as new to run")
 
@@ -64,6 +64,14 @@ def pytest_collection_modifyitems(config, items):
 
 
 
+def get_address(pytestconfig):
+    if pytestconfig.getoption('address'):
+        address=pytestconfig.getoption('address')
+    else:
+        port = pytestconfig.getoption('port')
+        ip   = pytestconfig.getoption('ip')
+        address = 'http://%s:%s' %(ip,port)
+    return address
 
 @pytest.fixture(scope="package")
 def chrome_driver(pytestconfig):
@@ -72,13 +80,8 @@ def chrome_driver(pytestconfig):
         # options.add_argument("--window-size=1920,1080")
         # options.add_argument("--start-maximized")
         options.add_argument('--headless')
+    address=get_address(pytestconfig)
 
-    if pytestconfig.getoption('address'):
-        address=pytestconfig.getoption('address')
-    else:
-        port = pytestconfig.getoption('port')
-        ip   = pytestconfig.getoption('ip')
-        address = 'http://%s:%s' %(ip,port)
     chrome_driver = webdriver.Chrome(options=options)
     # chrome_driver.set_network_conditions(offline=True,latency=5,throughput=500*1024)
     # print(address)
@@ -94,3 +97,12 @@ def sec(pytestconfig):
         sleep=0
     print(sleep)
     return sleep
+
+
+# @pytest.fixture(scope="package")
+# def port(pytestconfig):
+#     return pytestconfig.getoption('port')
+
+@pytest.fixture(scope="package")
+def address(pytestconfig):
+    return get_address(pytestconfig)